refactor: extract shared CardResultsPanel component
Unified header layout (card name, rarity pill, set code, card number,
price trend) was only on the home page. Deep link page had a different,
simpler layout that drifted out of sync.

- Created CardResultsPanel.svelte: owns unified header, two-column
  layout, pricing panel, lightbox, all derived values (pricing, image
  URL, header price/trend, rarity color)
- Slimmed home +page.svelte: removed ~80 lines of results markup/styles,
  now uses &lt;CardResultsPanel card={} set={} /&gt;
- Slimmed deep link +page.svelte: replaced old card-header layout with
  same CardResultsPanel, passes isLoading for skeleton state
- Both pages now render identical results UI from single source of truth
- Added to barrel export in components/index.ts
