+++ frontend/src/lib/components/index.ts
@@ -5,6 +5,7 @@ export { default as SetDropdownItem } from './SetDropdownItem.svelte';
 export { default as SetGroupHeader } from './SetGroupHeader.svelte';
 export { default as SearchForm } from './SearchForm.svelte';
 export { default as CardDetailPanel } from './CardDetailPanel.svelte';
+export { default as CardResultsPanel } from './CardResultsPanel.svelte';
 export { default as PricingPanel } from './PricingPanel.svelte';
 export { default as HeroPrice } from './HeroPrice.svelte';
 export { default as VariantPills } from './VariantPills.svelte';
