+++ frontend/src/routes/cards/[set_id]/[card_id]/+page.svelte
@@ -2,8 +2,7 @@
   import { onMount } from 'svelte';
   import { goto } from '$app/navigation';
   import { page } from '$app/stores';
-  import { SearchForm, CardDetailPanel, PricingPanel, CardVariantSelector, RecentLookups, SkeletonLoader } from '$lib/components';
-  import ImageLightbox from '$lib/components/ImageLightbox.svelte';
+  import { SearchForm, CardVariantSelector, RecentLookups, CardResultsPanel, SkeletonLoader } from '$lib/components';
   import { setsStore } from '$lib/stores/sets.svelte';
   import { cardsStore } from '$lib/stores/cards.svelte';
   import { pricingStore } from '$lib/stores/pricing.svelte';
@@ -21,24 +20,10 @@
   let selectedVariant = $state<any>(null);
   let isDeepLinkLoading = $state(true);
   let deepLinkError = $state<string | null>(null);
-  let lightboxUrl = $state<string | null>(null);
   let recentLookupsRef: ReturnType<typeof RecentLookups> | undefined = $state(undefined);
 
-  let currentPricing = $derived.by(() => {
-    const card = cardsStore.selectedCard;
-    const set = setsStore.selectedSet;
-    if (!card || !set) return null;
-    return pricingStore.priceData[`${set.id}_${card.id}`] || null;
-  });
+  let cardReady = $derived(!!setsStore.selectedSet && !!cardsStore.selectedCard);
 
-  let cardImageUrl = $derived.by(() => {
-    const card = cardsStore.selectedCard;
-    if (!card?.images || card.images.length === 0) return null;
-    const img = card.images[0];
-    return img.medium || img.small || img.large || null;
-  });
-
-  let cardName = $derived(cardsStore.selectedCard?.name ?? '');
   let pageTitle = $derived.by(() => {
     const card = cardsStore.selectedCard;
     const set = setsStore.selectedSet;
@@ -46,15 +31,11 @@
     return 'Loading... | PCPC';
   });
 
-  let cardReady = $derived(!!setsStore.selectedSet && !!cardsStore.selectedCard);
-
   function handlePriceFetched(info: { setId: string; cardId: string; name: string; imageUrl: string | null; setName: string; language: string }) {
     recentLookupsRef?.addLookup(info);
     goto(`/cards/${info.setId}/${info.cardId}`, { replaceState: true });
   }
 
-  function handleLightbox(url: string) { lightboxUrl = url; }
-  function closeLightbox() { lightboxUrl = null; }
   function handleVariantSelect(variant: any) { selectedVariant = variant; }
   function handleVariantConfirm(variant: any) { selectedVariant = variant; showVariantSelector = false; }
   function closeVariantSelector() { showVariantSelector = false; }
@@ -62,11 +43,8 @@
 
   /**
    * Find a set by ID, auto-switching language filter if needed.
-   * Detects language mismatch from the set ID pattern and switches
-   * to 'both' so deep links work regardless of saved preference.
    */
   async function findSetById(setId: string): Promise<import('$lib/types').PokemonSet | null> {
-    // Check current available sets
     let target = setsStore.availableSets.find((s) => s.id === setId) ?? null;
     if (target) return target;
 
@@ -77,7 +55,7 @@
       }
     }
 
-    // Not found — likely a language filter mismatch. Switch to 'both' and retry.
+    // Not found \u2014 likely a language filter mismatch. Switch to 'both' and retry.
     const isJpSet = setId.endsWith('_ja') || setId.includes('_ja_') || setId.includes('_ja-');
     const currentLang = setsStore.language;
 
@@ -157,41 +135,16 @@
     {/if}
 
     {#if isDeepLinkLoading && !cardReady}
-      <div class="results-container"><SkeletonLoader variant="pricing" /></div>
+      <div class="loading-container"><SkeletonLoader variant="pricing" /></div>
     {/if}
 
     {#if cardReady}
-      <div class="results-container">
-        <div class="results-layout">
-          <div class="results-sidebar">
-            <CardDetailPanel card={cardsStore.selectedCard} set={setsStore.selectedSet} imageUrl={cardImageUrl} onlightbox={handleLightbox} />
-          </div>
-          <div class="results-main">
-            <div class="card-header">
-              <h2 class="card-name">{cardsStore.selectedCard.name}</h2>
-              <p class="card-subtitle">
-                {setsStore.selectedSet.name}
-                {#if setsStore.selectedSet.code}<span class="sep">&#x00B7;</span> {setsStore.selectedSet.code.toUpperCase()}{/if}
-                {#if cardsStore.selectedCard.artist}<span class="sep">&#x00B7;</span> {cardsStore.selectedCard.artist}{/if}
-              </p>
-            </div>
-            {#if isDeepLinkLoading || pricingStore.isLoading}
-              <SkeletonLoader variant="pricing" />
-            {:else if currentPricing}
-              <PricingPanel pricing={currentPricing} />
-            {/if}
-          </div>
-        </div>
-      </div>
+      <CardResultsPanel card={cardsStore.selectedCard} set={setsStore.selectedSet} isLoading={isDeepLinkLoading} />
     {/if}
   </main>
 
   <CardVariantSelector variants={cardVariants} isVisible={showVariantSelector} onselect={handleVariantSelect} onconfirm={handleVariantConfirm} onclose={closeVariantSelector} />
 </div>
-
-{#if lightboxUrl}
-  <ImageLightbox imageUrl={lightboxUrl} altText="{cardName} - full size" onclose={closeLightbox} />
-{/if}
 
 <style>
   .pcpc-app { display: flex; flex-direction: column; min-height: 100vh; background-color: var(--bg-primary); color: var(--text-primary); position: relative; z-index: 1; }
@@ -212,15 +165,8 @@
   .error-close:hover { opacity: 0.7; background: none; }
   .back-link { background: none; border: 0.5px solid var(--color-error-text); color: var(--color-error-text); padding: 4px 10px; border-radius: var(--radius-badge); cursor: pointer; font-size: var(--fs-badge); white-space: nowrap; flex-shrink: 0; transition: opacity 0.15s ease; }
   .back-link:hover { opacity: 0.7; background: none; }
-  .results-container { background-color: var(--bg-container); border: 0.5px solid var(--border-subtle); border-radius: var(--radius-card); padding: 24px; box-shadow: var(--shadow-sm); position: relative; }
-  .results-container::before { content: ''; position: absolute; top: 0; left: 20px; right: 20px; height: 1px; background: linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.08) 30%, rgba(255, 255, 255, 0.12) 50%, rgba(255, 255, 255, 0.08) 70%, transparent); border-radius: 1px; pointer-events: none; }
-  .results-layout { display: flex; gap: var(--layout-gap, 24px); align-items: flex-start; }
-  .results-sidebar { flex-shrink: 0; position: sticky; top: 24px; align-self: flex-start; }
-  .results-main { flex: 1; min-width: 0; overflow: visible; position: relative; }
-  .card-header { margin-bottom: 4px; }
-  .card-name { margin: 0 0 4px 0; font-size: var(--fs-card-name); font-weight: 500; letter-spacing: -0.3px; color: var(--text-primary); }
-  .card-subtitle { margin: 0; font-size: var(--fs-body); color: var(--text-muted); }
-  .sep { color: var(--text-dim); margin: 0 2px; }
-  @media (max-width: 768px) { .header { padding: 12px 16px; } .app-title { font-size: 15px; } .main-content { padding: 16px; } .results-container { padding: 16px; } .results-layout { flex-direction: column; align-items: center; } .results-sidebar { position: static; } .card-header { text-align: center; } }
-  @media (max-width: 480px) { .header { padding: 10px 12px; } .app-title { font-size: 14px; } .main-content { padding: 12px; } .results-container { padding: 12px; border-radius: 8px; } }
+  .loading-container { background-color: var(--bg-container); border: 0.5px solid var(--border-subtle); border-radius: var(--radius-card); padding: 24px; box-shadow: var(--shadow-sm); }
+
+  @media (max-width: 768px) { .header { padding: 12px 16px; } .app-title { font-size: 15px; } .main-content { padding: 16px; } .loading-container { padding: 16px; } }
+  @media (max-width: 480px) { .header { padding: 10px 12px; } .app-title { font-size: 14px; } .main-content { padding: 12px; } .loading-container { padding: 12px; border-radius: 8px; } }
 </style>
